test(cli): isolate version-inspect parsing from mock.module leaks
rollback.test.ts replaces get-container-version for the whole process, so
the new inspect tests returned undefined when the CLI suite ran together.
